requirements.txt
curl_cffi==0.9.0
lxml==5.3.0
python-dotenv==1.0.1
pandas==2.2.3
XlsxWriter==3.2.0
requests==2.32.3